Refactor the notifications switch into its own component and disable it when the notification permission is not 'granted'
